# Only allocate a new (random) trace ID when trace ID parse fails rathe…