…r than all the time and prepare for immutable.

Noticed that we always allocate a fresh `TraceID` even when parsing one successfully. We should only do so in the case of a failure, since generating a random is very expensive. Most parses are successful. Around 4x speedup.

Also cleaned up delimiter parsing using @shengxil's pattern and prepared to make this an immutable class by deprecating setters / constructors and creating factories.

After
```
Benchmark                                                      Mode     Cnt     Score     Error   Units
TraceHeaderBenchmark.parse                                   sample  556594     0.555 ±   0.025   us/op
TraceHeaderBenchmark.parse:parse·p0.00                       sample             0.339             us/op
TraceHeaderBenchmark.parse:parse·p0.50                       sample             0.405             us/op
TraceHeaderBenchmark.parse:parse·p0.90                       sample             0.542             us/op
TraceHeaderBenchmark.parse:parse·p0.95                       sample             0.746             us/op
TraceHeaderBenchmark.parse:parse·p0.99                       sample             1.914             us/op
TraceHeaderBenchmark.parse:parse·p0.999                      sample            19.232             us/op
TraceHeaderBenchmark.parse:parse·p0.9999                     sample            52.097             us/op
TraceHeaderBenchmark.parse:parse·p1.00                       sample          1740.800             us/op
TraceHeaderBenchmark.parse:·gc.alloc.rate                    sample      15  1322.606 ±  78.021  MB/sec
TraceHeaderBenchmark.parse:·gc.alloc.rate.norm               sample      15   888.236 ±   0.024    B/op
TraceHeaderBenchmark.parse:·gc.churn.G1_Eden_Space           sample      15  1331.803 ± 119.214  MB/sec
TraceHeaderBenchmark.parse:·gc.churn.G1_Eden_Space.norm      sample      15   894.092 ±  54.121    B/op
TraceHeaderBenchmark.parse:·gc.churn.G1_Survivor_Space       sample      15     0.093 ±   0.014  MB/sec
TraceHeaderBenchmark.parse:·gc.churn.G1_Survivor_Space.norm  sample      15     0.062 ±   0.011    B/op
TraceHeaderBenchmark.parse:·gc.count                         sample      15   145.000            counts
TraceHeaderBenchmark.parse:·gc.time                          sample      15   128.000                ms
```

Before
```
Benchmark                                                      Mode     Cnt     Score    Error   Units
TraceHeaderBenchmark.parse                                   sample  437664     2.024 ±  0.209   us/op
TraceHeaderBenchmark.parse:parse·p0.00                       sample             0.498            us/op
TraceHeaderBenchmark.parse:parse·p0.50                       sample             0.800            us/op
TraceHeaderBenchmark.parse:parse·p0.90                       sample             0.934            us/op
TraceHeaderBenchmark.parse:parse·p0.95                       sample             1.170            us/op
TraceHeaderBenchmark.parse:parse·p0.99                       sample             4.496            us/op
TraceHeaderBenchmark.parse:parse·p0.999                      sample            40.021            us/op
TraceHeaderBenchmark.parse:parse·p0.9999                     sample          1761.280            us/op
TraceHeaderBenchmark.parse:parse·p1.00                       sample          3551.232            us/op
TraceHeaderBenchmark.parse:·gc.alloc.rate                    sample      15   627.784 ± 53.949  MB/sec
TraceHeaderBenchmark.parse:·gc.alloc.rate.norm               sample      15  1072.556 ±  0.061    B/op
TraceHeaderBenchmark.parse:·gc.churn.G1_Eden_Space           sample      15   631.644 ± 86.257  MB/sec
TraceHeaderBenchmark.parse:·gc.churn.G1_Eden_Space.norm      sample      15  1076.679 ± 92.875    B/op
TraceHeaderBenchmark.parse:·gc.churn.G1_Survivor_Space       sample      15     0.100 ±  0.017  MB/sec
TraceHeaderBenchmark.parse:·gc.churn.G1_Survivor_Space.norm  sample      15     0.172 ±  0.030    B/op
TraceHeaderBenchmark.parse:·gc.count                         sample      15    94.000           counts
TraceHeaderBenchmark.parse:·gc.time                          sample      15    95.000               ms
```


By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.