@anuraaga 2020-06-08
By the way, I find the order of methods to look weird when using factories. Filed an issue in aws sdk repo so we can consider the issue consistently https://github.com/aws/aws-sdk-java-v2/issues/1880