@willarmiros 2020-06-08
Also, do you think you could include the benchmarks you're running in the Java SDK [benchmark](https://github.com/aws/aws-xray-sdk-java/tree/master/aws-xray-recorder-sdk-benchmark) package for future reference?