@anuraaga 2020-06-08
Yeah I'm checking them all in, this one happened to be merged last time.

https://github.com/aws/aws-xray-sdk-java/blob/master/aws-xray-recorder-sdk-benchmark/tst/main/java/com/amazonaws/xray/entities/TraceHeaderBenchmark.java