@anuraaga 2020-06-08
Yup in 3.x will make this immutable, deprecated the setters in preparation.